Added some drink creation and alcohol content test cases

+++ Drunkr/src/drunkr/Ingredient.java
@@ -2,12 +2,12 @@ package drunkr;
 
 public class Ingredient {
 	public String name;
-	public int amount;
+	public double amount;
 	public String unit;
 
 	public double abv;
 	
-	public Ingredient(String name, int amount, String unit, double abv) {
+	public Ingredient(String name, double amount, String unit, double abv) {
 		this.name = name;
 		this.amount = amount;
 		this.unit = unit;
